requirements.txt
python-dotenv
pandas
xlrd
mysql-connector-python
streamlit
watchdog

